fix: add correct path to uuid
When uuid is used multiple times in a project with different version it is not installed in the root of the `node_modules` of the project.
When the `protractor.mock.js` is copied to a temp folder there is no relation to `ng-apimock` and thus no reference to `uuid` which is installed as a dependency from `ng-apimock`. This fix should solve the problem